fix: use BigInt for salary filter to prevent INT4 overflow
salaryMax values above ~2.1B (e.g. 3B VND × 1.2 = 3.6B) overflowed
PostgreSQL INT4 when passed as number. Wrap with BigInt() to match
the expectedSalaryMin/Max BigInt schema type.

+++ src/modules/cv-flip/cv-flip.service.ts
@@ -107,10 +107,10 @@ export class CvFlipService {
     const buildRangeConditions = (range: { min: number | undefined; max: number | undefined }): Prisma.UserProfileWhereInput[] => {
       const conditions: Prisma.UserProfileWhereInput[] = [];
       if (range.min !== undefined) {
-        conditions.push({ expectedSalaryMax: { gte: Math.floor(range.min) } });
+        conditions.push({ expectedSalaryMax: { gte: BigInt(Math.floor(range.min)) } });
       }
       if (range.max !== undefined) {
-        conditions.push({ expectedSalaryMin: { lte: Math.ceil(range.max) } });
+        conditions.push({ expectedSalaryMin: { lte: BigInt(Math.ceil(range.max)) } });
       }
       return conditions;
     };
